Refactored timestamp parsing functionality and date formats
