+++ content/dateformats.txt
@@ -1,0 +1,21 @@
+Source	Result	Normalized ISO
+2016-12-09	2016-12-09	2016-12-09
+2016-12-09 09:29	2016-12-09 09:29	2016-12-09 09:29
+2016-12-09 09:29:02,258	2016-12-09 09:29:02,258	2016-12-09 09:29:02,258
+2016-12-09 09:29:02,258211	2016-12-09 09:29:02,258211	2016-12-09 09:29:02,258211
+12/09/2016	12/09/2016	2016-12-09
+12.09.2016	12.09.2016	2016-12-09
+12-09-2016	12-09-2016	2016-12-09
+12-09-2016 09:29	12-09-2016 09:29	2016-12-09 09:29
+12-09-2016 09:29:02,258	12-09-2016 09:29:02,258	2016-12-09 09:29:02,258
+2016-12-09T09:29:02.258Z	2016-12-09T09:29:02.258Z	2016-12-09T09:29:02.258Z
+2016-12-09T09:29:02.258+01:00	2016-12-09T09:29:02.258+01:00	2016-12-09T09:29:02.258+01:00
+2016-12-09T09:29:02.258211+01:00	2016-12-09T09:29:02.258211+01:00	2016-12-09T09:29:02.258211+01:00
+2016-12-09T09:29:02.258211Z	2016-12-09T09:29:02.258211Z	2016-12-09T09:29:02.258211Z
+2016-12-09 09:29:02.258211 +01:00	2016-12-09 09:29:02.258211 +01:00	2016-02-09 09:29:02.258211 +01:00
+12/09/2016  09:29:02.25	12/09/2016  09:29:02.25	2016-12-09 09:29:02.25
+2016-12-09T0929	2016-12-09T09:29	2016-12-09 09:29:00
+09:29	09:29	09:29
+09:29:02	09:29:02	09:29:02
+09:29:02,258	09:29:02,258	09:29:02,258
+09:29:02,258211	09:29:02,258211	09:29:02,25821
